backend/requirements.txt
Django==4.2.20
djangorestframework==3.15.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1 